+++ Tests/KeyboardKitTests/Actions/KeyboardAction+Test.swift
@@ -19,7 +19,7 @@ extension KeyboardAction {
             .character(""),
             .command,
             .control,
-            .custom(name: ""),
+            .custom(named: ""),
             .dictation,
             .dismissKeyboard,
             .emoji(Emoji("")),
@@ -36,7 +36,7 @@ extension KeyboardAction {
             .keyboardType(.email),
             .keyboardType(.emojis),
             .keyboardType(.images),
-            .keyboardType(.custom("")),
+            .keyboardType(.custom(named: "")),
             
             .moveCursorBackward,
             .moveCursorForward,